requirements.txt
fastapi
uvicorn
pymongo
gradio
pydantic
pandas
python-dotenv
boto3